#1: Move SocketAddrSpec structure to shared module
