+++ src/shared/mod.rs
@@ -12,3 +12,4 @@ pub mod error;
 pub mod prelude;
 pub mod router_configuration;
 pub mod router_engine;
+pub mod socket_addr_spec;
